backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.12",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
    "argon2-cffi>=25.1.0",
    "openai>=2.7.2",
    "python-multipart>=0.0.20",
]
